Move transform config bits into own module
